libcore: Fix unchecked warning for Enum class
The comparable object was previously always assumed to be an Enum
without doing a check so a warning was output for it. It is actually
safe to do so as E extends Enum<E>, meaning the comparable object must
extend Enum<?>. Additionally, this fixes a build issue for JDK7 as the
compiler was complaining about the ordinal variable being private.

- PS1: change implemented
- PS2: commit message changed
- PS3: info about class casting issues removed

Change-Id: I07766b2756eaf09de2b054932bebc39dccce5731